requirements.txt
scipy
tqdm
numpy
joblib
pandas
ipython
librosa
TextGrid
sklearn
matplotlib
argparse
torchvision
kaldiio
PyYAML
